% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isop_login.R
\name{isop_login}
\alias{isop_login}
\title{Login to ISOP}
\usage{
isop_login(username = NULL, password = NULL, store = TRUE, reset = FALSE)
}
\arguments{
\item{username}{Username for ISOP login. See \code{Authentication} section for
details}

\item{password}{Password for ISOP login. See \code{Authentication} section for
details}

\item{store}{Store the \code{username} and \code{password} in the \code{.Renviron}. Default
is TRUE.}

\item{reset}{Resets the \code{username} and \code{password} values in the \code{.Renviron}. Default
is FALSE.}
}
\value{
defined global variable \code{isop_loginhash} with NDOP login hash cookie
}
\description{
This function is used to log in to ISOP. Account can be created at
\url{https://idm.nature.cz/idm/#/registration}. You can use the function
separately, otherwise the function will start automatically if a login is
required.
}
\section{Authentication}{

There are two option how to set login authentication:
\enumerate{
\item Set login credentials in the \code{.Renviron} file. This can be done
manually by editing the file, or just use the function without any paramatres
(\code{isop_login()}). In case that login credentials are not stored in
\code{.Renviron}, you will be promted, and the login will be than stored. R must
be restarted to take into account the changes in the \code{.Renviron} file.
\item Simply pass strings to each of the parameters. However, this is not
recommended, because the strings will be stored in \code{.Rhistory} file, which
can be easily, accidentally shared by user.
}
}

\examples{
# recommended
isop_login()

# not recommended
usr <- "isop_username"
pwd <- "isop_password"
isop_login(usr, pwd)

# using new credentials
isop_login(reset = T)
}
