% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndop_download.R
\name{ndop_download}
\alias{ndop_download}
\title{Download occurrence data from NDOP}
\usage{
ndop_download(taxon, locations = 0)
}
\arguments{
\item{taxon}{search by species or genus name}

\item{locations}{enables download raw localizations provided by NDOP in
ESRI shapefiles. The localizations are provided as points, lines and polygons
layers, and contains only geometries and lcalization id. Options: 1 -
download only localization (without table data); 2 - download localization
and table data; NULL - do not download localization (default)}
}
\value{
\code{data.frame} with table data or/and list of \code{sf} objects with
localizations (depend on \code{locations} parameter)
}
\description{
Download data from NDOP. Default data source is table data containing all
available informations about records. Optionally is also possible to
download raw localizations.
}
\section{Coordinates}{

Table data includes XY coordinates in S-JTSK (EPSG:5514). Source of some
records are polygons or lines, in that case XY coordinates in table are
centroids of source geometries. See \code{locations} parameter for downloading
raw geometries.
}

\examples{
# Exaple with table data
mr <- ndop_download("mantis religiosa")
head(mr)

require(sf)
sf_mr <- st_as_sf(x = mr,
               coords = c("X", "Y"),
               crs = st_crs(5514))

plot(sf_mr$geometry)

sf_mr$year <- as.numeric(substring(sf_mr$DATUM_OD,1,4))
hist(sf_mr$year,100)
plot(sf_mr["year"])

# Example with localization data

locs <- ndop_download("mantis religiosa", locations = 1)
plot(locs[[1]]$geometry)
}
