% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndop_list.R
\name{ndop_list}
\alias{ndop_list}
\title{Show promt value lists}
\usage{
ndop_list(list_type)
}
\arguments{
\item{list_type}{entry category of values: "species" for list of available
species values; "family" for families; and "group" for higher taxonomic
groups defined in NDOP}
}
\description{
Check all available values for NDOP queries of species, families, and groups.
}
\section{Note}{

Some of returned dataframes contains also \code{payload_val}, which is value used
in the \code{search_payload} list i.e. in POST requests during querying the server. For entering values in
\code{\link{ndop_download}} use values from \code{name_lat}.

Some group entries have multiple groups in one row (e.g. \verb{Collembola,  Protura, Diplura} with \code{payload_val} \code{119}), you can entry only one of them,
but is not possible to split the request, becaues the request use
the \code{payload_val}.
}

\examples{
# Exaple with table data
ndop_list("group")
}
