% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndop-package.R
\name{rndop-package}
\alias{rndop-package}
\alias{rndop}
\title{Access to the Data from The Species Occurrence Database of NCA CR (Czechia) in R}
\description{
A programmatic interface to the The Species Occurrence Database
of NCA CR (NDOP AOPK). The NDOP is Czech national species occurrence database
of NCA CR (Nature Conservation Agency of the Czech Republic;
\url{https://portal.nature.cz/nd/}). Database includes occurrence records both
from professionals (monitorings, inventarisations) and citizen scientists.
Use of the database is free and requires free registration
(\url{https://idm.nature.cz/idm/#/registration}). Data itself are published
under CC BY-SA 4.0 licence and should be referenced according to AOPK
citation rules
(\url{https://portal23.nature.cz/publik_syst/ctihtmlpage.php?what=4910}).
}
\author{
Oto Kaláb
}
