% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_search_payload.R
\name{set_search_payload}
\alias{set_search_payload}
\title{Generate a template for user search payload for POST request}
\usage{
set_search_payload(
  but_action = "Filtrovat",
  but_co = "rf",
  pagesizeX = 10000,
  aopkOrtho = 0,
  searchSelector = "ku",
  rfMesiceOd = "",
  rfMesiceDo = "",
  rfDatumOd = "1.1.0001",
  rfDatumDo = "1.1.9999",
  rfAkce = "",
  rfNalez = "",
  rfIdLokalizace = "",
  rfLokalizace = "",
  rfAutor = "",
  rfZdroj = "",
  rfProjekt = "",
  rfMetadata = "",
  rfZapsal = "",
  rfTaxon = "",
  rfKategorie = "",
  rfCeledi = "",
  rfKatastr = "",
  rfMZCHU = "",
  rfVZCHU = "",
  rfEVL = "",
  rfPO = "",
  rfZahrUzemi = "",
  rfKraj = "",
  rfPusobnost = "",
  rfSpatialX = "",
  rfSpatial = "",
  rfKvadrat = "",
  rfPresnost = "",
  rfNalezyCr = "",
  rfFotky = "",
  rfKO = "",
  rfSO = "",
  rfO = "",
  rfCcsCR = "",
  rfCcsEN = "",
  rfCcsVU = "",
  rfCcsNT = "",
  rfCcsEX = "",
  rfCcsDD = "",
  rfEVDII = "",
  rfEVDIV = "",
  rfEVDV = "",
  rfEVDI = "",
  rfBL = "",
  rfGL = "",
  rfWL = "",
  rf1143 = "",
  `text+inputObec` = "",
  parametryZakresu = "",
  existujeZakres = "",
  karta_id = "",
  karta_vztazne_id = "",
  idAkce = "",
  lpass = ""
)
}
\value{
\code{list}
}
\description{
Allows to generate a template of list of keys and values of user POST request for \code{\link{ndop_download}} function. This is useful for users who wants experiment, or are familiar with the content of the POST request sent to the.
}
\examples{
# Prepare search payload for 6175 KFME field
pld <- set_search_payload(rfKvadrat = 6175, num_rec_only = T)
}
