% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndop-package.R
\name{rndop-package}
\alias{rndop-package}
\alias{rndop}
\title{Interface for The Species Occurrence Database of NCA CR (Czechia)}
\description{
A programmatic interface to the The Species Occurrence Database
of NCA CR (NDOP AOPK). The NDOP is Czech national species occurrence database
of NCA CR (Nature Conservation Agency of the Czech Republic; <https://portal.
nature.cz/nd/>). Database includes occurrence records both from professionals
(monitorings, inventarisations) and citizen scientists. Use of the database
is free and requires free registration
(\url{https://idm.nature.cz/idm/#/egistration}). Data itselfs are published
under CC BY-SA 4.0 licence and should be referenced according to AOPK
citation rules
(\url{https://portal.nature.cz/publik_syst/ctihtmlpage.php?what=4910}).
}
\author{
Oto Kaláb
}
